--- src/components/pages/MakeSticker/PicSticker/index.jsx
@@ -1,13 +1,0 @@
-import picselect from "/src/assets/svgs/PicSelect.svg";
-
-const PicSticker = () => {
-  return (
-    <div className="bg-syscolor-SystemPurple1">
-      <button>
-        <img src={picselect} />
-      </button>
-    </div>
-  );
-};
-
-export default PicSticker;
